+++ components/aboutCard/navBar/navText/NavText.tsx
